+++ Assets/LTH/LTH_Scripts/Items/TestCode/ItemType.cs.meta
