--- Assets/LTH/Scripts/Items/TestCode/ItemType.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace LTH
-{
-    public enum ItemType
-    {
-      Cigarette, // 담배
-      Cuffs, // 수갑
-      Dial, // 다이얼
-      Cellphone, // 휴대폰
-    }
-}
